# Packet benchmarks and optimizations