--- a/libraries/earendil_crypt/Cargo.toml
+++ b/libraries/earendil_crypt/Cargo.toml
@@ -11,9 +11,10 @@ arrayref = "0.3.7"
 base32 = "0.4.0"
 blake3 = "1.5.0"
 bytes = "1.5.0"
-ed25519-compact = "2.0.4"
+
 rand = "0.8.5"
 serde = { version = "1.0.188", features = ["derive"] }
 stdcode = "0.1.14"
 thiserror = "1.0.49"
 base64 = "0.21.5"
+ed25519-consensus = "2.1.0"