--- a/libraries/earendil_packet/Cargo.toml
+++ b/libraries/earendil_packet/Cargo.toml
@@ -22,6 +22,7 @@ bs58 = "0.5.0"
 serde-big-array = "0.5.1"
 earendil_crypt = { path = "../earendil_crypt" }
 base64 = "0.21.5"
+x25519-dalek = {version="2.0.0", features=["reusable_secrets"]}
 
 [dev-dependencies]
 criterion = "0.3"