--- a/libraries/earendil_packet/Cargo.toml
+++ b/libraries/earendil_packet/Cargo.toml
@@ -12,7 +12,7 @@ bytemuck = { version = "1.14.0", features = ["derive", "min_const_generics"] }
 bytes = { version = "1.5.0", features = ["serde"] }
 chacha20 = "0.9.1"
 chacha20poly1305 = "0.10.1"
-ed25519-compact = "2.0.4"
+
 rand = "0.8.5"
 thiserror = "1.0.49"
 serde = { version = "1.0.188", features = ["derive"] }
@@ -30,3 +30,6 @@ criterion = "0.3"
 [[bench]]
 name = "benchmark"
 harness = false
+
+[profile.release]
+debug = true  # Include debug symbols in release builds