--- a/libraries/earendil_packet/Cargo.toml
+++ b/libraries/earendil_packet/Cargo.toml
@@ -30,4 +30,3 @@ criterion = "0.3"
 [[bench]]
 name = "benchmark"
 harness = false
-