--- a/libraries/earendil_topology/Cargo.toml
+++ b/libraries/earendil_topology/Cargo.toml
@@ -23,3 +23,4 @@ rand = "0.8.5"
 indexmap = { version = "1.9.3", features = ["serde"] }
 thiserror= "1.0.49"
 earendil_crypt={path="../earendil_crypt"}
+