- `criterion` benchmarks are added to `earendil_packet`
- moved x25519 to `x25519-dalek` and ed25519 to `ed25519-consensus`
- removed pointless source authentication for N2R messages

All this together shows a significant improvement in performance, especially for single-hop routes:

```
generate an OnionSecret time:   [38.610 ns 38.621 ns 38.636 ns]                                     
                        change: [-99.934% -99.932% -99.931%] (p = 0.00 < 0.05)
                        Performance has improved.

1-hop RawPacket construction                                                                            
                        time:   [298.87 µs 299.14 µs 299.48 µs]
                        change: [-65.115% -63.156% -61.351%] (p = 0.00 < 0.05)
                        Performance has improved.

1-hop ReplyBlock construction                                                                            
                        time:   [320.77 µs 320.81 µs 320.85 µs]
                        change: [-52.486% -52.441% -52.396%] (p = 0.00 < 0.05)
                        Performance has improved.

4-hop RawPacket construction                                                                            
                        time:   [709.19 µs 709.40 µs 709.76 µs]
                        change: [-38.414% -38.384% -38.352%] (p = 0.00 < 0.05)
                        Performance has improved.

4-hop ReplyBlock construction                                                                            
                        time:   [740.20 µs 740.73 µs 741.34 µs]
                        change: [-36.529% -36.461% -36.387%] (p = 0.00 < 0.05)
                        Performance has improved.

Benchmarking 8-hop RawPacket construction: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 6.5s, enable flat sampling, or reduce sample count to 60.
8-hop RawPacket construction                                                                             
                        time:   [1.2833 ms 1.2841 ms 1.2849 ms]
                        change: [-28.369% -28.322% -28.278%] (p = 0.00 < 0.05)
                        Performance has improved.

Benchmarking 8-hop ReplyBlock construction: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 6.5s, enable flat sampling, or reduce sample count to 60.
8-hop ReplyBlock construction                                                                             
                        time:   [1.2743 ms 1.2748 ms 1.2754 ms]
                        change: [-30.000% -29.913% -29.798%] (p = 0.00 < 0.05)
                        Performance has improved.
 ```